use HttpClient directly
