+++ dot-net/ReliableDownloader/FileDownloadService.cs
@@ -2,7 +2,7 @@
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 
-namespace ReliableDownloader;
+namespace Accurx.ReliableDownloader;
 
 internal record FileDownloadServiceSettings
 {
@@ -12,34 +12,25 @@ internal record FileDownloadServiceSettings
     public required string DestinationFilePath { get; init; }
 }
 
-internal sealed class FileDownloadService : IHostedService
+internal sealed class FileDownloadService(
+    ILogger<FileDownloadService> logger,
+    IOptions<FileDownloadServiceSettings> downloadSettings,
+    IFileDownloader fileDownloader)
+    : IHostedService
 {
-    private readonly ILogger<FileDownloadService> _logger;
-    private readonly IOptions<FileDownloadServiceSettings> _downloadSettings;
-    private readonly IFileDownloader _fileDownloader;
-
-    public FileDownloadService(
-        ILogger<FileDownloadService> logger,
-        IOptions<FileDownloadServiceSettings> downloadSettings,
-        IFileDownloader fileDownloader)
-    {
-        _logger = logger;
-        _downloadSettings = downloadSettings;
-        _fileDownloader = fileDownloader;
-    }
-
     public async Task StartAsync(CancellationToken cancellationToken)
     {
-        var didDownloadSuccessfully = await _fileDownloader.TryDownloadFile(
-            _downloadSettings.Value.SourceUrl,
-            Path.Combine(Directory.GetCurrentDirectory(), _downloadSettings.Value.DestinationFilePath), cancellationToken);
+        var didDownloadSuccessfully = await fileDownloader.TryDownloadFile(
+            downloadSettings.Value.SourceUrl,
+            Path.Combine(Directory.GetCurrentDirectory(), downloadSettings.Value.DestinationFilePath),
+            cancellationToken);
 
-        _logger.LogInformation("File download ended! Success: {DownloadSuccessful}", didDownloadSuccessfully);
+        logger.LogInformation("File download ended! Success: {DownloadSuccessful}", didDownloadSuccessfully);
     }
 
     public Task StopAsync(CancellationToken cancellationToken)
     {
-        _logger.LogInformation("FileDownloadService is stopping.");
+        logger.LogInformation("FileDownloadService is stopping.");
         return Task.CompletedTask;
     }
 }
