--- dot-net/ReliableDownloader/FileProgress.cs
@@ -1,7 +1,0 @@
-﻿namespace ReliableDownloader;
-
-internal record FileProgress(
-    long? TotalFileSize,
-    long TotalBytesDownloaded,
-    double? ProgressPercent,
-    TimeSpan? EstimatedRemaining);
