--- dot-net/ReliableDownloader/ISystemCalls.cs
@@ -1,6 +1,0 @@
-﻿namespace ReliableDownloader;
-
-internal interface ISystemCalls
-{
-    Stream FileOpen(string localFilePath, FileMode openOrCreate, FileAccess readWrite);
-}
